requirements.txt
torch==2.0.1
torchaudio==2.0.2
numpy==1.24.3
scipy==1.11.2
librosa==0.10.0
soundfile==0.12.1
matplotlib==3.7.2
tensorboard==2.13.0
tqdm==4.66.1
pyyaml==6.0
flask==2.3.3
werkzeug==2.3.7
